+++ src/js/helpers/autologin.js
@@ -4,10 +4,11 @@ module.exports = function autologin( loginMethod, form ) {
   var visualStatusReport = document.getElementById( 'loading__status' );
   var url = window.location;
   var params = new URLSearchParams( url.search );
+  var loginMethodDisplayName = NLX.displayNames.loginMethod ? NLX.displayNames.loginMethod : loginMethod;
   var newLocation;
 
   form.willRedirect = true;
-  visualStatusReport.textContent = 'Attempting auto-login with ' + loginMethod;
+  visualStatusReport.textContent = 'Attempting auto-login with ' + loginMethodDisplayName;
 
   params.set( 'sso', 'true' );
   params.set( 'connection', loginMethod );