+++ src/js/decorators/check-if-maintenance-mode.js
@@ -1,9 +1,9 @@
 var ui = require( 'helpers/ui' );
 
 module.exports = function checkIfMaintenanceMode( banner ) {
-  var maintenanceMode = NLX.features.maintenanceMode;
+  var maintenance_mode = NLX.features.maintenance_mode;
 
-  if ( maintenanceMode === 'true' ) {
+  if ( maintenance_mode === 'true' ) {
     ui.show( banner );
   }
 };